+++ prism_home/scripts/init.lua
@@ -1,0 +1,3 @@
+main.on_bootstrap(function ()
+    logger.info("test")
+end)