+++ src/Squid.Prism.Server.Core/Interfaces/Services/ISquidPrismServiceProvider.cs
@@ -1,0 +1,9 @@
+namespace Squid.Prism.Server.Core.Interfaces.Services;
+
+public interface ISquidPrismServiceProvider
+{
+    TService GetService<TService>();
+
+    object? GetService(Type serviceType);
+
+}